added scanBasePackages

+++ src/main/java/interfaces/IUserRepo.java
@@ -3,6 +3,6 @@ package interfaces;
 import db.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IUserRepo extends JpaRepository <User, Integer>
+public interface IUserRepo extends JpaRepository <User, Long>
 {
 }
